+++ app/models/data_source.rb
@@ -8,10 +8,21 @@ class DataSource < ActiveRecord::Base
     class AbstractTable < ActiveRecord::Base
       self.abstract_class = true
 
-      class_attribute :defined_at
+      class_attribute :defined_at, :schema_name
 
       def self.cache_key
         "#{name.underscore}-#{defined_at.strftime("%Y%m%d%H%M%S")}"
+      end
+
+      def access_data_source
+        return yield if schema_name == "_"
+        original_search_path = connection.schema_search_path
+        begin
+          connection.schema_search_path = schema_name
+          yield
+        ensure
+          connection.schema_search_path = original_search_path
+        end
       end
     end
   end
@@ -61,26 +72,50 @@ class DataSource < ActiveRecord::Base
     base_class
   end
 
-  def source_table_names
-    RequestStore.store["data_source_source_table_names_#{id}"] ||= source_base_class.connection.tables.reject {|table_name| ignored_table_patterns.match(table_name) }
+  def source_tables
+    tables = case source_base_class.connection
+      when ActiveRecord::ConnectionAdapters::PostgreSQLAdapter, ActiveRecord::ConnectionAdapters::RedshiftAdapter
+        source_base_class.connection.query(<<-SQL, 'SCHEMA')
+          SELECT schemaname, tablename
+          FROM (
+            SELECT schemaname, tablename FROM pg_tables WHERE schemaname = ANY (current_schemas(false))
+            UNION
+            SELECT schemaname, viewname AS tablename FROM pg_views WHERE schemaname = ANY (current_schemas(false))
+          ) tables
+          ORDER BY schemaname, tablename;
+        SQL
+      else
+        source_base_class.connection.tables.map {|table_name| ["_", table_name] }
+    end
+    tables.reject {|_, table_name| ignored_table_patterns.match(table_name) }
   end
 
-  def source_table_class(table_name, table_names=source_table_names)
+  def cached_source_tables
+    RequestStore.store["data_source_source_tables_#{id}"] ||= source_tables
+  end
+
+  def source_schema_names
+    cached_source_tables.map {|table| table[0] }.uniq.sort
+  end
+
+  def source_table_class(table_name, tables=source_tables)
     return if ignored_table_patterns.match(table_name)
     table_class_name = source_table_class_name(table_name)
     return DynamicTable.const_get(table_class_name) if DynamicTable.const_defined?(table_class_name)
-    return nil unless table_names.include?(table_name)
+    schema_name, _ = tables.find {|_, table| table == table_name }
+    return nil unless schema_name
     table_class = Class.new(source_base_class)
     table_class.table_name = table_name
+    table_class.schema_name = schema_name
     DynamicTable.const_set(table_class_name, table_class)
     table_class.defined_at = Time.now
     table_class
   end
 
   def source_table_classes
-    table_names = source_table_names
-    table_names.map do |table_name|
-      source_table_class(table_name, table_names)
+    tables = source_tables
+    tables.map do |_, table_name|
+      source_table_class(table_name, tables)
     end
   end